Add Laravel Reverb (WebSocket)

+++ app/Services/Auth/AuthService.php
@@ -3,12 +3,14 @@
 namespace App\Services\Auth;
 
 use App\Enums\Role\RoleType;
+use App\Events\User\UserLoggedOutEverywhere;
 use App\Models\User;
 use App\Services\DTO\Auth\LoginDTO;
 use App\Services\DTO\Auth\RegisterDTO;
 use App\Services\DTO\Response\SimpleResponse;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Laravel\Sanctum\PersonalAccessToken;
 
 final class AuthService implements AuthServiceInterface
 {
@@ -18,6 +20,7 @@ final class AuthService implements AuthServiceInterface
             'email' => $dto->getEmail(),
             'password' => $dto->getPassword()
         ], $dto->isRemember());
+
         if (!$auth) {
             return new SimpleResponse(false, [
                 'message' => 'Неверный логин или пароль'
@@ -26,7 +29,7 @@ final class AuthService implements AuthServiceInterface
 
         /** @var User $user */
         $user ??= Auth::user();
-        $user->tokens()->delete();
+
         $token = $user->createToken('auth_token')->plainTextToken;
 
         return new SimpleResponse(true, [
@@ -45,15 +48,30 @@ final class AuthService implements AuthServiceInterface
             'role_id' => RoleType::User->value,
             'department_id' => $dto->getDepartmentId()
         ]);
+
         $user->markEmailAsVerified();
+
         return $this->login($dto, $user);
     }
 
-    public function logout(User $user): SimpleResponse
+    public function logout(User $user, ?PersonalAccessToken $currentToken = null, bool $logoutFromAllDevices = false): SimpleResponse
     {
-        $user->tokens()->delete();
+        if ($logoutFromAllDevices) {
+            $user->tokens()->delete();
+
+            broadcast(new UserLoggedOutEverywhere($user->id))->toOthers();
+
+            return new SimpleResponse(true, [
+                'message' => 'Успешно произведен выход со всех устройств'
+            ]);
+        }
+
+        if ($currentToken !== null) {
+            $currentToken->delete();
+        }
+
         return new SimpleResponse(true, [
-            'message' => 'Успешно произведен выход со всех устройств'
+            'message' => 'Успешно произведен выход на текущем устройстве'
         ]);
     }
 }
